Standardize DASH tests location. (#277)
- Moved all existing tests under test/test-cases
- Updated docker files to add/mount new location
- Updated documentation.
