+++ dash-pipeline/README-dash-workflows.md
@@ -75,13 +75,13 @@ You also have to build `sai` in order for the bmv2 "forwarding pipeline config" 
 make run-switch                   # console 1
 make init-switch                  # console 2
 make run-saithrift-client-bash    # console 2
-...   
+...
 root@chris-z4:/tests-dev/saithrift# scapy
 >>> p=Ether()/IP()/UDP()
 >>> sendp(p, iface='veth1')
 .
 Sent 1 packets.
->>> 
+>>>
 ```
 
 ## Use-Case II - Developing P4 Code + libsai config (C++)
@@ -122,7 +122,7 @@ Once you have stable P4 code, `libsai` and a saithrift client/server framework, 
 # Make Target Summary
 The tables below summarize the most important `make` targets for easy reference. You can click on a link to jump to further explanations. Not all make targets are shown. See the [Makefile](Makefile) to learn more.
 
-Dockerfile build targets are separately described in [README-dash-docker](README-dash-docker.md) since they are mainly for infrastructure and generally not part of day-to-day code and test-case development. The one exception is the [docker-saithrift-client](#build-saithrift-client-docker-image) target. 
+Dockerfile build targets are separately described in [README-dash-docker](README-dash-docker.md) since they are mainly for infrastructure and generally not part of day-to-day code and test-case development. The one exception is the [docker-saithrift-client](#build-saithrift-client-docker-image) target.
 ## Make "ALL" Targets
 | Target(s)              | Description                                                                  |
 | ---------------------- | --------------------------------------------------|
@@ -131,7 +131,7 @@ Dockerfile build targets are separately described in [README-dash-docker](README
 | [kill-all](#stop-containers)             | Stops all running containers                      |
 | [run-all-tests](#run-all-tests)          | Run all tests under `dash-pipeline/tests`         |
 
-## Build Artifacts 
+## Build Artifacts
 | Target(s)              | Description                                                                  |
 | ---------------------- | --------------------------------------------------|
 | [p4](#compile-p4-code) <br>[p4-clean](#compile-p4-code) | Compiles P4 code and produces both bmv2 and P4Info `.json` files.<br>Delete p4 artifacts |
@@ -151,14 +151,14 @@ Dockerfile build targets are separately described in [README-dash-docker](README
 | Target(s)              | Description                                                                  |
 | ---------------------- | --------------------------------------------------|
 | [run-libsai-test](run-libsai-c-tests)     | Run tests under [tests/libsai](tests/libsai)       |
-| [run-saithrift-ptftests](#run-saithrift-client-ptf-tests) | Run PTF tests under [tests/saithrift/ptf](tests/libsai/ptf) using tests built into [docker-saithrift-client](#build-saithrift-client-docker-image) image
+| [run-saithrift-ptftests](#run-saithrift-client-ptf-tests) | Run PTF tests under [test/test-cases/functional](../test/test-cases/functional) using tests built into [docker-saithrift-client](#build-saithrift-client-docker-image) image
 | [run-saithrift-pytests](#run-saithrift-client-pytests) | Run Pytests under [tests/saithrift/pytest](tests/libsai/pytest) using tests built into [docker-saithrift-client](#build-saithrift-client-docker-image) image
 |[run-saithrift-client-tests](#run-saithrift-client-tests) | Run all saithrift tests |
-| [run-saithrift-dev-ptftests](#run-saithrift-client-ptf-tests) <br> [run-saithrift-dev-pytests](#run-saithrift-client-dev-pytests) <br> [run-saithrift-client-dev-tests](#run-saithrift-client-dev-tests) | Like the three targets above. above, but run tests from host directory `tests/saithrift` instead of tests built into the `saithrift-client` container for faster test-case development code/test cycles.
+| [run-saithrift-dev-ptftests](#run-saithrift-client-ptf-tests) <br> [run-saithrift-dev-pytests](#run-saithrift-client-dev-pytests) <br> [run-saithrift-client-dev-tests](#run-saithrift-client-dev-tests) | Like the three targets above. above, but run tests from host directory [test/test-cases/functional](../test/test-cases/functional) instead of tests built into the `saithrift-client` container for faster test-case development code/test cycles.
 
 # Detailed DASH Behavioral Model Build Workflow
 
-This explains the various build steps in more details. The CI pipeline does most of these steps as well. All filenames and directories mentioned in the sections below are relative to the `dash-pipeline` directory (containing this README) unless otherwise specified. 
+This explains the various build steps in more details. The CI pipeline does most of these steps as well. All filenames and directories mentioned in the sections below are relative to the `dash-pipeline` directory (containing this README) unless otherwise specified.
 
 The workflows described here are primarily driven by a [Makefile](Makefile) and are suitable for a variety of use-cases:
 * Manual execution by developers - edit, build, test; commit and push to GitHub
@@ -305,7 +305,7 @@ Switch is initialized.
 ### Use wireshark to decode P4Runtime messages in the SAI-P4RT adaptor
 >**Hint:** You can monitor P4Runtime messages using Wireshark or similar. Select interface `lo`, filter on `tcp.port==9559`. Right-click on a captured packet and select "Decode as..." and configure port 9559 to decode as HTTP2 (old versions of Wireshark might lack this choice).
 ## Run saithrift-server
->**Note:** the bmv2 switch must be running, see 
+>**Note:** the bmv2 switch must be running, see
 When this server is launched, it will establish a P4Runtime session (behind the scenes) to the running `bmv2` switch daemon . The thrift server listens on port `9092` for Thrift messages carrying SAI rpc commands. These commands are dispatched the the SAI library handlers. These handlers translate them into corresponding P4Runtime RPC commands and are sent to the bmv2 daemon onto a socket at standard P4Runtime port `9559`.
 ```
 make run-saithrift-server
@@ -342,15 +342,15 @@ To run all "Production" tests which use the saithrift interface, execute the fol
 ```
 make run-saithrift-client-tests
 ```
-This will launch a saithrift-client docker container and execute tests under `tests/saithrift`, including:
+This will launch a saithrift-client docker container and execute tests under `test/test-cases/functional`, including:
 * Pytests under `tests/saithrift/pytest`
-* PTF Tests under `tests/saithrift/PTF`
+* PTF Tests under `test/test-cases/functional`
 ### Run saithrift-client PTF tests
 To run all PTF tests which use the saithrift interface, execute the following. You must have the bmv2 switch and saithrift-server running.
 ```
 make run-saithrift-client-ptftests
 ```
-This will launch a saithrift-client docker container and execute tests under `tests/saithrift/ptf`.
+This will launch a saithrift-client docker container and execute tests under `test/test-cases/functional`.
 ### Run saithrift-client Pytests
 To run all Pytests which use the saithrift interface, execute the following. You must have the bmv2 switch and saithrift-server running.
 ```
