+++ dash-pipeline/dockerfiles/Dockerfile.saithrift-client
@@ -9,13 +9,13 @@ the saithrift client and test programs for DASH."
 #
 
 # Copy distro PTF submodule and tools from SAI repo
-ADD SAI/SAI/test/ptf /SAI/test/ptf
+ADD dash-pipeline/SAI/SAI/test/ptf /SAI/test/ptf
 # Install PTF test framework & test-cases from SAI repo
-ADD SAI/SAI/ptf /SAI/ptf/
-# Copy thrift python distro 
-ADD SAI/rpc/thrift-0.11.0.tar.gz /
+ADD dash-pipeline/SAI/SAI/ptf /SAI/ptf/
+# Copy thrift python distro
+ADD dash-pipeline/SAI/rpc/thrift-0.11.0.tar.gz /
 # Copy autogenerated saithrift library built from SAI headers for DASH dataplane
-ADD SAI/rpc/saithrift-0.9.tar.gz /
+ADD dash-pipeline/SAI/rpc/saithrift-0.9.tar.gz /
 # Install the python libraries
 RUN cd /saithrift-0.9 && \
     sudo python3 setup.py install && \
@@ -29,8 +29,11 @@ RUN cd /saithrift-0.9 && \
     sudo python3 setup.py install
 
 # Copy distro test-cases into container, making it standalone (doesn't need host FS contents).
-# For dev, host directories can be mounted to another container directory to "see" tests in development    
-ADD tests/ tests/
+# For dev, host directories can be mounted to another container directory to "see" tests in development
+ADD test/test-cases/ tests/
+
+# Copy dash-pipeline tests into container
+ADD dash-pipeline/tests/ ci-tests/
 
 WORKDIR /
