+++ dash-pipeline/tests/saithrift/ptf/README.md
@@ -2,3 +2,5 @@
 **TODO - Placeholder**
 
 This directory can contain PTF tests for DASH bmv2. These would supplement standard tests under [SAI/ptf](../../SAI/ptf)
+
+CI-ready DASH tests are moved to [`test/test-cases/functional/ptf`](../../../../test/test-cases/functional/ptf/)
