+++ test/test-cases/README.md
@@ -7,16 +7,3 @@ This contains a hierarchical set of directories containing DASH test cases organ
 | --- | --- |
 | [functional](./functional/) | Tests to verify essential functionality using low-rate traffic (SAI PTF).
 | [scale](./scale/) | Tests with high-rate traffic and complex configuration to verify scaling real-world scenarios.
-
-## functional
-
-| Folder/File | Description |
-| --- | --- |
-| [`saidashvnet.py`](./functional/saidashvnet.py) | VNET-to-VNET test cases
-
-
-## scale
-
-| Folder/File | Description |
-| --- | --- |
-| [vnet2vnet](scale/vnet2vnet/README.md) | DASH vnet2vnet Tests
