+++ test/test-cases/functional/README.md
@@ -1,0 +1,11 @@
+# Functional Test Cases
+Tests to verify essential functionality using low-rate traffic (SAI PTF).
+
+## Contents
+
+| Folder/File | Description |
+| --- | --- |
+| [`saidashacl.py`](./ptf/saidashacl.py) | DASH ACL test cases
+| [`saidashvnet_sanity.py`](./ptf/saidashvnet_sanity.py) | Basic positive VNET-to-VNET test cases designed for running on BMv2
+| [`saidashvnet.py`](./ptf/saidashvnet.py) | More VNET-to-VNET test cases designed to provide a full feature coverage
+
