+++ test/test-cases/run-functional-tests.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
 # To be run inside saithrift-client container, assumes SAI repo portions exist under /SAI directory
 
-ptf --test-dir functional --pypath /SAI/ptf --interface 0@veth1 --interface 1@veth3 $@
+ptf \
+    --test-dir functional \
+    --pypath /SAI/ptf \
+    --interface 0@veth1 --interface 1@veth3 \
+    --test-case-timeout=1800 \
+    --test-params="bmv2=True" \
+    $@
