+++ test/test-cases/scale/README.md
@@ -1,0 +1,8 @@
+# Scale Test Cases
+Tests with high-rate traffic and complex configuration to verify scaling real-world scenarios.
+
+## scale
+
+| Folder/File | Description |
+| --- | --- |
+| [vnet2vnet](vnet2vnet/README.md) | Scale DASH vnet2vnet Tests
